#28 create readme and clean up files
Rename the import folder to "in" to reduce confusion
Explain what python scripts are doing
Remove the SQL files in import/locations, as they are not needed anymore and may lead to problems, because they don't have deterministic uuids.
Remove zip files in input directory as they are redundant and partly even faulty (e.g. A discrepancy between German and International infrastructure data [0.5] #26 not fixed in zip)